extracted team_member_row component into separate file
